move last (commented) webdriver test to upload.spec.ts, for later conversion + remove last references to webdriver
